pyproject.toml
[tool.poetry]
name = "dreamcatalog"
version = "0.1.0"
description = "Enhance product photos using AI"
authors = ["niy <niy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
streamlit = "^1.20.0"
removebg = "^0.4"
openai = "^0.27.2"
pillow = "^9.3.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.8.4"
rembg = "^2.0.30"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
